Make review latency summary card optional
For the Animations dashboard, we'd like to be able to display the per-person
code review cards without the code review latency summary card. This patch gives
cz-review-latency-dash an argument to indicate whether the summary card is shown
or not, with the blink.json config including the summary and animations-ave.json
config not including it.